Refactored the rest of the html parsing functions into the parsing module
And added new tests for the functions in the module
