+++ edx_dl/parsing.py
@@ -5,8 +5,13 @@ Parsing and extraction functions
 """
 import re
 
+from bs4 import BeautifulSoup as BeautifulSoup_
+
 from datetime import timedelta, datetime
-from .common import Unit
+from .common import Course, Section, SubSection, Unit
+
+# Force use of bs4 with html5lib
+BeautifulSoup = lambda page: BeautifulSoup_(page, 'html5lib')
 
 def edx_json2srt(o):
     """
@@ -85,3 +90,62 @@ def extract_units_from_html(page, BASE_URL):
                               resources_urls=resources_urls))
 
     return units
+
+
+def extract_courses_from_html(page, BASE_URL):
+    """
+    Extracts courses (Course) from the html page
+    """
+    soup = BeautifulSoup(page)
+    courses_soup = soup.find_all('article', 'course')
+    courses = []
+    for course_soup in courses_soup:
+        course_id = None
+        course_name = course_soup.h3.text.strip()
+        course_url = None
+        course_state = 'Not yet'
+        try:
+            # started courses include the course link in the href attribute
+            course_url = BASE_URL + course_soup.a['href']
+            if course_url.endswith('info') or course_url.endswith('info/'):
+                course_state = 'Started'
+            # The id of a course in edX is composed by the path
+            # {organization}/{course_number}/{course_run]
+            course_id = course_soup.a['href'][9:-5]
+        except KeyError:
+            pass
+        courses.append(Course(id=course_id,
+                              name=course_name,
+                              url=course_url,
+                              state=course_state))
+    return courses
+
+
+def extract_sections_from_html(page, BASE_URL):
+    """
+    Extract sections (Section->SubSection) from the html page
+    """
+    def _make_url(section_soup):  # FIXME: Extract from here and test
+        return BASE_URL + section_soup.ul.find('a')['href']
+
+    def _get_section_name(section_soup):  # FIXME: Extract from here and test
+        return section_soup.h3.a.string.strip()
+
+    def _make_subsections(section_soup):
+        subsections_soup = section_soup.ul.find_all("li")
+        # FIXME correct extraction of subsection.name (unicode)
+        subsections = [SubSection(position=i,
+                                  url=BASE_URL + s.a['href'],
+                                  name=s.p.string)
+                       for i, s in enumerate(subsections_soup, 1)]
+        return subsections
+
+    soup = BeautifulSoup(page)
+    sections_soup = soup.find_all('div', attrs={'class': 'chapter'})
+
+    sections = [Section(position=i,
+                        name=_get_section_name(section_soup),
+                        url=_make_url(section_soup),
+                        subsections=_make_subsections(section_soup))
+                for i, section_soup in enumerate(sections_soup, 1)]
+    return sections
